+++ lua/mini/git.lua
@@ -1305,7 +1305,7 @@ H.setup_repo_watch = function(buf_id, repo)
     local on_change = vim.schedule_wrap(function() H.on_repo_change(repo) end)
     local watch = function(_, filename, _)
       -- Ignore temporary changes
-      if vim.endswith(filename, 'lock') then return end
+      if vim.endswith(filename or '', 'lock') then return end
 
       -- Debounce to not overload during incremental staging (like in script)
       timer:stop()