+++ tests/test_extra.lua
@@ -2608,7 +2608,7 @@ local setup_keymaps = function()
 end
 
 T['pickers']['keymaps()']['works'] = function()
-  child.set_size(30, 80)
+  child.set_size(31, 80)
   setup_keymaps()
 
   child.lua_notify('_G.return_item = MiniExtra.pickers.keymaps()')
@@ -2617,7 +2617,7 @@ T['pickers']['keymaps()']['works'] = function()
 
   -- Should have proper preview
   type_keys('<Tab>')
-  child.expect_screenshot()
+  if child.fn.has('nvim-0.12') == 1 then child.expect_screenshot() end
 
   -- Should properly choose by executing LHS keys
   type_keys('<CR>')
@@ -3173,6 +3173,204 @@ T['pickers']['lsp()']['validates arguments'] = function()
 
   validate({}, '`pickers%.lsp` needs an explicit scope')
   validate({ scope = '1' }, '`pickers%.lsp`.*"scope".*"1".*one of')
+end
+
+T['pickers']['manpages()'] = new_set({
+  hooks = {
+    pre_case = function()
+      if child.fn.has('nvim-0.10') == 0 then
+        expect.error(function() child.lua('MiniExtra.pickers.manpages()') end, '`manpages` picker needs Neovim>=0%.10')
+        MiniTest.skip('`manpages` picker needs Neovim>=0.10')
+      end
+
+      -- Mock `vim.loop.spawn` for `MiniPick.builtin.cli()`
+      mock_spawn()
+
+      -- Mock `vim.system` for preview
+      child.cmd('luafile tests/mock-system/vim-system.lua')
+
+      -- Mock `vim.fn.executable` for a robust testing
+      child.lua([[
+        _G.has_col = false
+        local executable_orig = vim.fn.executable
+        vim.fn.executable = function(expr)
+          if expr == 'col' then return _G.has_col and 1 or 0 end
+          return executable_orig(expr)
+        end
+      ]])
+    end,
+  },
+})
+
+local pick_manpages = forward_lua_notify('MiniExtra.pickers.manpages')
+
+local mock_man_list = function()
+  mock_cli_return({
+    -- "Regular" output
+    'st (1)               - simple terminal',
+    'IO::Socket::IP (3perl) - Family-neutral IP socket supporting both IPv4 and IPv6',
+
+    -- "Exotic" output (like Void Linux, OpenBSD)
+    -- - Several commands separated by comma
+    'alacritty, Alacritty (1) - A fast, cross-platform, OpenGL terminal emulator.',
+    -- - Several sections separated by comma
+    'afterstep_faq (1, 2)',
+    -- - No space before `(section)`
+    'alacritty-msg(1) - Send messages to Alacritty.',
+    -- - Unexpected characters in `(section)`
+    'amd64_iopl(2/amd64)',
+  })
+end
+
+T['pickers']['manpages()']['works'] = function()
+  child.set_size(20, 70)
+  mock_man_list()
+
+  child.fn.setenv('PATH', '/home,/home/user')
+  child.fn.setenv('MANPATH', '/home,/home/manpage')
+  child.lua_notify('_G.return_item = MiniExtra.pickers.manpages()')
+  validate_picker_name('Manpages')
+  child.expect_screenshot()
+
+  local cwd = child.fn.getcwd()
+  local spawn_env = { 'MANWIDTH=999', 'PATH=/home,/home/user', 'MANPATH=/home,/home/manpage' }
+  eq(get_spawn_log(), { { executable = 'man', options = { args = { '-k', '.' }, cwd = cwd, env = spawn_env } } })
+  clear_spawn_log()
+  clear_process_log()
+
+  -- Should preview with correct `vim.system` usage
+  child.lua([[
+    _G.system_queue = {
+      { stdout = 'st (1) - Manual Page' },
+      { stdout = 'IO::Socket::IP (3perl) - Manual Page' },
+      { stdout = 'alacritty (1) - Manual Page' },
+      { stdout = 'afterstep_faq (1) - Manual Page' },
+      { stdout = 'alacritty-msg(1) - Manual Page' },
+      { stdout = 'amd64_iopl(2/amd64) - Manual Page' },
+    }
+  ]])
+
+  local preview_win_id = child.lua_get('MiniPick.get_picker_state().windows.main')
+  local preview_width = child.api.nvim_win_get_width(preview_win_id)
+
+  local validate_manpage_buf = function(buf_id, ref_name)
+    -- Should set proper options in preview buffer (for syntax highlight)
+    local scope = child.fn.has('nvim-0.11') == 1 and 'local' or nil
+    eq(child.api.nvim_get_option_value('modified', { scope = scope, buf = buf_id }), false)
+    eq(child.api.nvim_get_option_value('filetype', { scope = scope, buf = buf_id }), 'man')
+
+    local name = child.api.nvim_buf_get_name(buf_id)
+    if ref_name ~= nil then
+      eq(name, ref_name)
+    else
+      expect.match(name, 'minipick://%d+/preview')
+    end
+  end
+
+  local validate_preview = function(ref_cmd)
+    local ref_log = {
+      { 'vim.system', { cmd = ref_cmd, opts = { env = { MANWIDTH = preview_width, MANPAGER = 'cat' } } } },
+      { 'wait', {} },
+    }
+    eq(child.lua_get('_G.system_log'), ref_log)
+    child.lua('_G.system_log = {}')
+
+    -- NOTE: Should not name buffer to avoid name conflict after `choose`
+    validate_manpage_buf(child.api.nvim_win_get_buf(preview_win_id), nil)
+  end
+
+  type_keys('<Tab>')
+
+  -- - Should use syntax `man <section> <command>`
+  validate_preview({ 'man', '1', 'st' })
+  child.expect_screenshot()
+
+  -- - Should extract valid section value (see `:Man man`)
+  type_keys('<C-n>')
+  validate_preview({ 'man', '3perl', 'IO::Socket::IP' })
+
+  -- - Should use first command if there are several
+  type_keys('<C-n>')
+  validate_preview({ 'man', '1', 'alacritty' })
+
+  -- - Should use first section if there are several
+  type_keys('<C-n>')
+  validate_preview({ 'man', '1', 'afterstep_faq' })
+
+  -- - Should handle no space before `(section)`
+  type_keys('<C-n>')
+  validate_preview({ 'man', '1', 'alacritty-msg' })
+
+  -- - Should handle non-digits in `(section)`
+  type_keys('<C-n>')
+  validate_preview({ 'man', '2', 'amd64_iopl' })
+
+  -- Should properly choose by showing manpage in target window (not split)
+  child.lua([[_G.system_queue = { { stdout = 'amd64_iopl(2/amd64) - Full page' } }]])
+  type_keys('<CR>')
+  child.expect_screenshot()
+  validate_buf_name(0, 'man://amd64_iopl(2)')
+
+  local target_win_width = child.api.nvim_win_get_width(0)
+  local ref_env = { MANPAGER = 'cat', MANWIDTH = target_win_width }
+  local choose_log = {
+    { 'vim.system', { cmd = { 'man', '2', 'amd64_iopl' }, opts = { env = ref_env } } },
+    { 'wait', {} },
+  }
+  eq(child.lua_get('_G.system_log'), choose_log)
+
+  -- Should return chosen value
+  eq(child.lua_get('_G.return_item'), 'amd64_iopl(2/amd64)')
+
+  -- Should work without set up 'mini.pick'
+  child.mini_unload('pick')
+  mock_man_list()
+  pick_manpages()
+  validate_active_picker()
+end
+
+T['pickers']['manpages()']['can choose in split'] = function()
+  mock_man_list()
+  pick_manpages()
+
+  child.lua([[_G.system_queue = { { stdout = 'st (1) - Full page' } }]])
+  type_keys('<C-v>')
+  child.expect_screenshot()
+end
+
+T['pickers']['manpages()']['can choose proper manpager'] = function()
+  child.lua('_G.has_col = true')
+  mock_man_list()
+  pick_manpages()
+
+  child.lua([[_G.system_queue = {
+    { stdout = 'st (1) - Full page' },
+    { stdout = 'st (1) - Full page' },
+  }]])
+  local preview_win_id = child.lua_get('MiniPick.get_picker_state().windows.main')
+  local preview_width = child.api.nvim_win_get_width(preview_win_id)
+
+  type_keys('<Tab>')
+  local ref_env = { MANWIDTH = preview_width, MANPAGER = 'col -bx' }
+  local ref_log = {
+    { 'vim.system', { cmd = { 'man', '1', 'st' }, opts = { env = ref_env } } },
+    { 'wait', {} },
+  }
+  eq(child.lua_get('_G.system_log'), ref_log)
+  child.lua('_G.system_log = {}')
+
+  type_keys('<CR>')
+  ref_log[1][2].opts.env.MANWIDTH = child.api.nvim_win_get_width(0)
+  eq(child.lua_get('_G.system_log'), ref_log)
+end
+
+T['pickers']['manpages()']['respects `opts`'] = function()
+  pick_manpages({}, { source = { name = 'My name' } })
+  validate_picker_name('My name')
+end
+
+T['pickers']['manpages()']['respects global source options'] = function()
+  validate_global_source_options(pick_manpages, false, false)
 end
 
 T['pickers']['marks()'] = new_set()